requirements.txt
fastapi~=0.108.0
pydantic~=2.5.3
motor~=3.3.2
Faker~=22.1.0
python-dotenv~=1.0.0